requirements.txt
flask
python-dotenv
openai
aiohttp
requests
langchain
langchain-openai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pytz